Move DB functions to 'database.scala'
